correcting switch in adminMenu

+++ src/main/java/se/yrgo/spring/client/LibraryApplication.java
@@ -106,7 +106,7 @@ public class LibraryApplication {
     private static void spacer(String x) {
         System.out.println(x.repeat(10));
     }
-    
+
     private static void signUp(UserService user, Set<String> ids, UniqueIdGenerator idGenerator, Scanner input)
             throws UserNotFoundException {
         String email;
@@ -402,14 +402,14 @@ public class LibraryApplication {
                     """);
             choice = input.nextLine();
             switch (choice) {
-                case "2" -> {
+                case "1" -> {
                     System.out.println("Lägg till författare.");
                     System.out.println("Namn: ");
                     String authorName = input.nextLine();
 
                     author.addAuthor(idGenerator.generateUniqueId(ids), authorName);
                 }
-                case "3" -> {
+                case "2" -> {
                     System.out.println("Skriv namn på författare du vill ta bort:");
                     String authorName = input.nextLine();
 
@@ -432,15 +432,17 @@ public class LibraryApplication {
 
     // Adam
     // Contributed to creating method manageLoans
+    // A method for the admin menu to handle loans. Includes updating and removing
+    // loans.
     private static void manageLoans(LoanService loan, Scanner input) {
         String choice;
         boolean loanMenu = true;
         System.out.println("Lån");
         while (loanMenu) {
             List<Loan> loans = loan.getAllLoans();
-            if (loans.isEmpty()){
+            if (loans.isEmpty()) {
                 System.out.println("Det finns inga lån för tillfället.");
-            }else{
+            } else {
                 loans.forEach(System.out::println);
                 spacer("-");
             }
@@ -478,6 +480,8 @@ public class LibraryApplication {
         }
     }
 
+    // A method for the admin menu that handles books. Includes adding and removing
+    // books.
     private static void manageBooks(AuthorService author, BookService book, Set<String> ids,
             UniqueIdGenerator idGenerator, Scanner input) throws BookNotFoundException {
         String choice;
@@ -486,8 +490,12 @@ public class LibraryApplication {
             System.out.println("Böcker");
             spacer("-");
             List<Book> books = book.getEntireCatalogue();
-            books.forEach(System.out::println);
-            spacer("-");
+            if (books.isEmpty()) {
+                System.out.println("Det finns inga böcker för tillfället.");
+            } else {
+                books.forEach(System.out::println);
+                spacer("-");
+            }
             System.out.printf("""
                     1. Lägg till bok
                     2. Ta bort bok
@@ -526,10 +534,9 @@ public class LibraryApplication {
                         System.out.println("Denna bok finns redan i systemet, ISBN: " + isbn);
                     } else {
                         book.registerNewBook(isbn, title, authors);
-                        System.out.println("Bok registrerad. \n ISBN: " + isbn + "\nTitel: " + title + "\nFöfattare: "
-                                + authorName);
+                        System.out.println("\nBok registrerad.");
                     }
-                    cleanScreen();
+                    //cleanScreen();
                 }
                 case "2" -> {
                     System.out.println("Ange bokens ISBN för att radera: ");
@@ -555,14 +562,20 @@ public class LibraryApplication {
         }
     }
 
+    // A method in the admin menu that handles users. Includes update and deletion
+    // of user.
     private static void manageUsers(UserService user, Scanner input) {
         String choice;
         boolean userMenu = true;
         System.out.println("Användare:\n");
         while (userMenu) {
             List<User> users = user.getAllUsers();
-            users.forEach(System.out::println);
-            spacer("-");
+            if (users.isEmpty()) {
+                System.out.println("Det finns inga användare för tillfället.");
+            } else {
+                users.forEach(System.out::println);
+                spacer("-");
+            }
             System.out.printf("""
                     1. Uppdatera användare
                     2. Ta bort användare
@@ -661,8 +674,16 @@ public class LibraryApplication {
                         cleanScreen();
                         break;
                     }
-                    user.deleteUser(id);
-                    System.out.println("Användare borttagen.");
+                    List<User> existingUsers = user.getAllUsers();
+                    boolean userExists = existingUsers.stream()
+                            .anyMatch(b -> b.getUserId().equals(id));
+
+                    if (!userExists) {
+                        System.out.println("Nånting gick fel vid radering av användare med ID: " + id);
+                    } else {
+                        user.deleteUser(id);
+                        System.out.println("Användare borttagen.");
+                    }
                 }
                 case "0" -> {
                     userMenu = false;
